--- a/package.json
+++ b/package.json
@@ -99,7 +99,6 @@
     "path": "0.12.7",
     "prop-types": "15.7.2",
     "query-string": "7.0.1",
-    "react-copy-to-clipboard": "5.0.4",
     "react-markdown": "7.1.1",
     "rehype-raw": "6.1.0",
     "rheostat": "4.1.1",