--- a/src/components/copy-to-clipboard.tsx
+++ b/src/components/copy-to-clipboard.tsx
@@ -1,4 +1,4 @@
-import { useState, useCallback, useRef, useEffect } from 'react';
+import { useState, useCallback, useRef, useEffect, memo } from 'react';
 import { sleep } from 'timing-functions';
 
 import { Button, ButtonProps } from './button';
@@ -18,51 +18,53 @@ type Props = {
   afterCopy?: string;
 };
 
-const CopyToClipboard = ({
-  textToCopy,
-  beforeCopy = 'Copy',
-  afterCopy = 'Copied',
-  ...props
-}: Props & ButtonProps) => {
-  const [copied, setCopied] = useState(false);
-  const copyPromise = useRef<Promise<void> | null>(null);
+const CopyToClipboard = memo(
+  ({
+    textToCopy,
+    beforeCopy = 'Copy',
+    afterCopy = 'Copied',
+    ...props
+  }: Props & ButtonProps) => {
+    const [copied, setCopied] = useState(false);
+    const copyPromise = useRef<Promise<void> | null>(null);
 
-  useEffect(() => {
-    setCopied(false);
-    return () => {
-      // Clear any potential ongoing promise on changing value or unmounting
-      copyPromise.current = null;
-    };
-  }, [textToCopy]);
+    useEffect(() => {
+      setCopied(false);
+      return () => {
+        // Clear any potential ongoing promise on changing value or unmounting
+        copyPromise.current = null;
+      };
+    }, [textToCopy]);
 
-  const handleClick = useCallback(() => {
-    const p = navigator.clipboard
-      .writeText(textToCopy)
-      .then(() => {
-        if (copyPromise.current === p) {
-          // Display copied for 10 seconds
-          setCopied(true);
-          return sleep(10000);
-        }
-        return null;
-      })
-      .then(() => {
-        if (copyPromise.current === p) {
-          setCopied(false);
-        }
-      });
-    copyPromise.current = p;
-  }, [textToCopy]);
+    const handleClick = useCallback(() => {
+      const p = navigator.clipboard
+        .writeText(textToCopy)
+        .then(() => {
+          if (copyPromise.current === p) {
+            // Display copied for 10 seconds
+            setCopied(true);
+            return sleep(10000);
+          }
+          return null;
+        })
+        .then(() => {
+          if (copyPromise.current === p) {
+            setCopied(false);
+          }
+        });
+      copyPromise.current = p;
+    }, [textToCopy]);
 
-  if (!('clipboard' in navigator) || !('writeText' in navigator.clipboard)) {
-    return null;
+    if (!('clipboard' in navigator) || !('writeText' in navigator.clipboard)) {
+      return null;
+    }
+
+    return (
+      <Button {...props} onClick={handleClick}>
+        {copied ? afterCopy : beforeCopy}
+      </Button>
+    );
   }
-
-  return (
-    <Button {...props} onClick={handleClick}>
-      {copied ? afterCopy : beforeCopy}
-    </Button>
-  );
-};
+);
 
 export default CopyToClipboard;