--- a/src/components/index.ts
+++ b/src/components/index.ts
@@ -8,6 +8,7 @@ export { default as ButtonModal } from './button-modal';
 export { default as Card } from './card';
 export { default as Chip } from './chip';
 export { default as CodeBlock } from './code-block';
+export { default as CopyToClipboard } from './copy-to-clipboard';
 export { DataList, DataListWithLoader } from './data-list';
 export { DataTable, DataTableWithLoader } from './data-table';
 export { default as DisplayMenu } from './display-menu';