--- a/src/styles/components/sequence.scss
+++ b/src/styles/components/sequence.scss
@@ -3,29 +3,53 @@
 
 .sequence-container {
   background-color: rgba($colour-platinum, 0.7);
+
   .action-bar {
     justify-content: flex-start;
   }
+
   .action-bar,
   .info-list {
     margin: $global-margin * 0.5;
   }
 
   .sequence {
-    display: flex;
-    align-items: flex-start;
+    font-family: $font-family-monospace;
+    overflow-wrap: anywhere;
 
-    &__sequence {
-      font-family: $font-family-monospace;
-      display: flex;
-      flex-wrap: wrap;
+    counter-reset: sequence;
 
-      &__chunk {
-        margin: $global-margin * 0.5;
-        &__counter {
-          text-anchor: end;
-          user-select: none;
-        }
+    &__chunk {
+      display: inline-block;
+      margin: $global-margin 0 $global-margin $global-margin;
+      width: 10ch;
+      position: relative;
+      white-space: nowrap;
+
+      &:not(:last-child)::before {
+        content: counter(sequence);
+        counter-increment: sequence 10;
+
+        position: absolute;
+        top: -0.8em;
+        right: 0;
+        opacity: 0.5;
+        font-weight: bold;
+      }
+
+      &::after {
+        content: '';
+
+        position: absolute;
+        bottom: 0;
+        left: 0;
+        right: 0;
+
+        transform-origin: left;
+        /* squeeze to 10%, so that each shadow will have the width of 1ch */
+        transform: scaleX(0.1);
+
+        box-shadow: var(--box-shadow);
       }
     }
   }