… src/components/__tests__/__snapshots__/copy-to-clipboard.spec.tsx.snap
… src/components/__tests__/__snapshots__/info-list.spec.tsx.snap
… src/components/__tests__/__snapshots__/sequence-chunk.spec.jsx.snap
… src/components/__tests__/__snapshots__/sequence.spec.jsx.snap
… src/components/__tests__/__snapshots__/sequence.spec.tsx.snap
… src/components/__tests__/copy-to-clipboard.spec.jsx
… src/components/__tests__/sequence-chunk.spec.jsx
… src/components/__tests__/sequence.spec.jsx
… src/components/copy-to-clipboard.jsx
… src/components/sequence-chunk.jsx
… src/mock-data/sample-data.js
… src/mock-data/sequence-data.ts
… yarn.lock