## Purpose
Fix for sequence highlights in Firefox https://www.ebi.ac.uk/panda/jira/browse/TRM-26297 + optimisations

## Approach
- Updated everything I touched to TypeScript
- Removed a library dependency to do the copy button, handled in the codebase
- Did an InfoList change to not render in no content (updated tests too). Was asked for by Dan in a uniprot-website PR, but I messed up my branches and ended up here
- Sequence:
  - Switched from rendering everything in SVG to having everything in HTML and reducing as much as possible the number of DOM nodes in the document
  - Numbering of chunks done in CSS pseudo-elements and using CSS counters
  - Highlighting done now in CSS pseudo-elements by (ab)using CSS shadows: squish the pseudo-element to a tenth of its size, and render plenty of shadows to represent highlights for every character. Coloring and positioning done by using CSS custom properties
  - Reduced the number of React components to use by using one memoised SequenceChunks components
  - Generate the highlight CSS code once and only once, only on the first highlight render, in order to avoid unecessary calculation, but optimise the follow-up changes if the user plays with the highlights.

### Before/after benchmark:
Done on a isolated storybook (so might have a performance and memory overhead), using the longest sequence I could find A0A5A9P0L4 (45,354 AA)

  | Before | Before, all highlights | After | After, all highlights
-- | -- | -- | -- | --
number of DOM nodes in `.sequence` | 18145 | 69868 | 4536 | 4536
JS heap size (memory usage) | 156MB | 231MB | 146MB | 149MB
Onload event | 3s 800ms | N/A | 3s 160ms | N/A
Blocking time on 1st highlight activation | 3s 500ms | N/A | 700ms | N/A
Blocking time on 2nd highlight activation | 3s 500ms | N/A | 300ms | N/A

Did multiple runs, and results were sensibly the same, so I haven't put more values in the table

This is mainly to be taken as a before/after comparison that things have improved, not as absolute performance values as it would need to be isolated even more, outside of storybook.

## Testing
- Updated snapshots and test
- Manual tests in Chrome, Firefox, and Safari
- Lowered test coverage as now:
  - styling is done in CSS would probably require a visual test
  - Copying is not handled by a library but ourselves, so slightly more code in our codebase and can't really test it because it's not supported in Jest (had to mock the clipboard)

## Stories
InfoList
Sequence

## Checklist
- [ ] My PR is scoped properly, and “does one thing only”
- [x] I have reviewed my own code
- [x] I have checked that linting checks pass and type safety is respected
- [ ] I have checked that tests pass and coverage has at least improved, and if not explained the reasons why
- [x] Were all the edited/created files in TypeScript? If not, please list the files you left in JavaScript and the reason to not update them.
- [ ] For the stories you created/updated, are the props modifiables through knobs?